+++ verdeciel/Location.Bank.swift
@@ -89,7 +89,7 @@ class LocationBank : Location
 		
 		var i = 0
 		for mesh in structure.childNodes {
-			mesh.eulerAngles.y = Float(degToRad((CGFloat(i) * 0.10)))
+			mesh.eulerAngles.y = (degToRad((CGFloat(i) * 0.10)))
 			i += 1
 		}
